+++ CHANGES.rst
@@ -5,6 +5,7 @@ UNRELEASED
 ----------
 
 * Support django.db.models.JSONField
+* List choices of choice fields
 
 
 Version 1.1 (2021-03-02)